Rename coverage.py to coverage_provider.py so we can use the coverage.py library to measure code coverage.
